requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
auto-py-to-exe==2.42.0
bcrypt==4.1.3
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
cryptography==42.0.7
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
Eel==0.16.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.5.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
googletrans==4.0.0rc1
greenlet==3.0.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.6.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.23.2
hyperframe==5.2.0
idna==3.7
inputimeout==1.0.4
install==1.3.5
instaloader==4.11
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.25.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
looseversion==1.3.0
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
paramiko==3.4.0
parso==0.8.4
pathlib==1.0.1
pefile==2023.2.7
platformdirs==4.2.0
prompt_toolkit==3.0.45
proto-plus==1.23.0
protobuf==4.25.3
proxy_tools==0.1.0
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.2
pydantic_core==2.18.3
pydot==2.0.0
Pygments==2.18.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.2
PyNaCl==1.5.0
pyparsing==3.1.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pythonnet==3.0.3
pytz==2024.1
pywebview==4.4.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.3
rdflib==7.0.0
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
safetensors==0.4.3
scipy==1.13.1
selenium==4.20.0
setuptools==69.1.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.27
stack-data==0.6.3
sympy==1.12.1
tabula==1.0.5
tbb==2021.12.0
textblob==0.18.0.post0
tokenizers==0.19.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
traits==6.3.2
transformers==4.41.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.0
wcwidth==0.2.13
Werkzeug==3.0.1
whichcraft==0.6.1
wsproto==1.2.0
xlrd==0.7.1
XlsxWriter==3.2.0
xlwt==0.7.2
yarl==1.9.4
zope.event==5.0
zope.interface==6.2
